+++ src/pages/api/subscribe.ts
@@ -1,12 +1,10 @@
 import { NextApiResponse } from 'next';
-import Mailchimp from 'mailchimp-api-v3';
 
 import { ISubscribeRequest } from '@/models/api/request/subcribe';
 import { ISubscribeResponse } from '@/models/api/response/subscribe';
-import { IMailchimpError } from '@/models/mailchimp';
 import { validateEmail } from '@/utils/validators';
-
-const MAILCHIMP_MEMBER_EXISTS_MESSAGE = 'Member Exists';
+import MailchimpAPI, { IMailchimpError } from '@/utils/mailchimp-api';
+import { MailchimpErrors } from '@/models/mailchimp';
 
 export default async function handler(req: ISubscribeRequest, res: NextApiResponse<ISubscribeResponse>) {
 	if (req.method === 'POST') {
@@ -20,11 +18,9 @@ export default async function handler(req: ISubscribeRequest, res: NextApiRespon
 			return;
 		}
 
-		const mailchimp = new Mailchimp(process.env.MAILCHIMP_API_KEY);
-
 		// Try to store the email in mailchimp
 		try {
-			await mailchimp.post(`/lists/${process.env.MAILCHIMP_LIST_ID}/members`, {
+			await MailchimpAPI.post(`/lists/${process.env.MAILCHIMP_LIST_ID}/members`, {
 				email_address: req.body.email,
 				status: 'subscribed',
 				tags: [process.env.MAILCHIMP_APPROVAL_LANDING_SUB_TAG],
@@ -37,8 +33,10 @@ export default async function handler(req: ISubscribeRequest, res: NextApiRespon
 
 			return;
 		} catch (e) {
+			const mailchimpError = e as IMailchimpError<{ readonly title?: string }>;
+
 			// The API request above might fail due to duplicate subscription try - in this case return successful response
-			const isDuplicateError = (e as IMailchimpError).title === MAILCHIMP_MEMBER_EXISTS_MESSAGE;
+			const isDuplicateError = mailchimpError.response?.data.title === MailchimpErrors.MemberExists;
 
 			if (isDuplicateError) {
 				res.status(200).send({
@@ -51,7 +49,7 @@ export default async function handler(req: ISubscribeRequest, res: NextApiRespon
 
 			res.status(500).send({
 				success: false,
-				message: e as string,
+				message: 'Failed to subscribe',
 			});
 
 			return;